requirements.txt
google-api-python-client==2.86.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
openai==1.3.5
python-dotenv==1.0.0
pytube==15.0.0
rich==13.4.2
typer==0.9.0
requests==2.32.0
streamlit==1.33.0
pytest==8.2.0
